refactor: Move Open API 2/3 JSON schemas to `specs.openapi` package
