+++ src/schemathesis/specs/openapi/definitions.py
